benchmark against most upvoted solutions

+++ 7kyu/herons-formula-1/src/lib.rs
@@ -1,6 +1,6 @@
 //! <https://www.codewars.com/kata/57aa218e72292d98d500240f/train/rust>
 
-pub fn heron(a: f64, b: f64, c: f64) -> f64 {
-    let s = (a + b + c) / 2.;
-    (s * (s - a) * (s - b) * (s - c)).sqrt()
+pub fn heron(a: u32, b: u32, c: u32) -> f64 {
+    let s = (a + b + c) as f64 / 2.;
+    (s * (s - a as f64) * (s - b as f64) * (s - c as f64)).sqrt()
 }
